Added script attaching in the editor on objects
Now you can attach scripts from game dll to game objects in scene.

Changed the resolution of the window to 1280x720 so it is a bit bigger.

The name of the scripts in game dll must be unique or else there could be some bugs. This is because a map is used by script name. This shouldn't be a problem because the script names are the same as the type name using the macro in the game project. So it should raise an error.

In the ui, the extra option popup isn't showing when right clicking a script node.